Improved the method of quoting strings for displaying them in messages:
-   All instances of using `\'"${parameter//\'/\'\\\'\'}"\'` were
    replaced with `"${parameter@Q}"`.
-   All instances of using `ls -1d "${array[@]}"` were replaced with
    `printf '%s\n' "${array[@]@Q}"`.

Improved handling of removal of trailing forward slashes when passing
the `<path>[/]` value to `-f, --edl-save-files-relative` and
`-F, --edl-save-files-absolute`, and when passing a directory as a
non-option argument:

-   Now, not one, but all trailing forward slashes are removed.
-   If the path consists of forward slashes only (one or more of them),
    a single forward slash representing the root directory is kept.

Improved formatting of the built-in help message.
Improved phrasing regarding the default value of
`-f, --edl-save-files-relative` and `-F, --edl-save-files-absolute` in
the man page.
Fixed further inaccuracies and formatting shortcomings in the script's
built-in help message, the man page, the README, and the CHANGELOG
files.

Removed a superfluous `IFS=' '` from the `getopt` line.